requirements.txt
pydantic
requests
Authlib
pycryptodomex